requirements.txt
fastapi
uvicorn
google-api-python-client
earthengine-api
python-dotenv
